digitalizacion_aceria/backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4..4
pyserial==3.5
python-dotenv==1.0.0